requirements.txt
# Core AI Framework
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-google-genai==2.0.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Web Interface
gradio==4.44.0

# Utilities
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Validation
email-validator==2.2.0
phonenumbers==8.13.47

# Logging
colorlog==6.8.2
